+++ packages/browser-wallet/src/wallet-api/wallet-api.ts
@@ -1,5 +1,5 @@
 import { InjectedMessageHandler, createEventTypeFilter, MessageType, MessageStatusWrapper } from '@messaging';
-import { Transaction } from '@concordium/web-sdk';
+import { Transaction, VerifiablePresentationV1, VerificationRequestV1 } from '@concordium/web-sdk';
 import {
     AccountAddress,
     AccountTransactionSignature,
@@ -334,6 +334,44 @@ class WalletApi extends EventEmitter implements IWalletApi {
         }
 
         return VerifiablePresentation.fromString(res.result);
+    }
+
+    public async requestVerifiablePresentationV1(request: VerificationRequestV1.Type) {
+        // Extract statements to be used in UI representation and validation
+        const statementsConverted = request.subjectClaims.map(({ statements, issuers, source }) => ({
+            statement: statements,
+            source,
+            idQualifier: {
+                type: 'cred',
+                issuers: issuers.map((issuer) => issuer.index),
+            },
+        }));
+
+        // Create Hex placeholder for validation pass
+        const challenge = 'A'.repeat(64);
+
+        // Helper object with old payload structure for validation reuse
+        // Can be used in this way, because new subject claims is same as old ones
+        // And if mapping of new ones will fail, this will also validate proper payload structure
+        const representationObj = {
+            statements: stringify(statementsConverted),
+            challenge,
+        };
+
+        const res = await this.messageHandler.sendMessage<MessageStatusWrapper<VerifiablePresentationV1.JSON>>(
+            MessageType.Web3IdProof,
+            {
+                // We have to stringify the statements because they can contain bigints
+                verificationRequestV1: stringify(request),
+                ...representationObj,
+            }
+        );
+
+        if (!res.success) {
+            throw new Error(res.message);
+        }
+
+        return VerifiablePresentationV1.fromJSON(res.result);
     }
 }